requirements.txt
psutil==7.0.0
pyyaml==6.0.2
requests==2.32.3
python-dotenv==1.1.0
